+++ crates/cli/src/lib.rs
@@ -11,9 +11,6 @@ use std::{collections::HashMap, io::Write, str::FromStr};
 use tokio::{io::AsyncReadExt, net::TcpListener, runtime::Handle};
 use tracing::metadata::Level;
 
-#[cfg(feature = "python")]
-mod python;
-
 /// stacrs: A command-line interface for the SpatioTemporal Asset Catalog (STAC)
 #[derive(Debug, Parser)]
 pub struct Stacrs {
@@ -372,7 +369,7 @@ impl Stacrs {
                     }
                     #[cfg(not(feature = "pgstac"))]
                     {
-                        return Err(anyhow!("stacrs is not compiled with pgstac support"));
+                        Err(anyhow!("stacrs is not compiled with pgstac support"))
                     }
                 } else {
                     let backend = stac_server::MemoryBackend::new();